Move Chrome and Firefox ETW flags out into different files.
