+++ samply/src/windows/firefox_etw_flags.rs
@@ -1,0 +1,5 @@
+// From https://searchfox.org/mozilla-central/rev/0e7394a77cdbe1df5e04a1d4171d6da67b57fa17/mozglue/baseprofiler/public/BaseProfilerMarkersPrerequisites.h#355-360
+pub const PHASE_INSTANT: u8 = 0;
+pub const PHASE_INTERVAL: u8 = 1;
+pub const PHASE_INTERVAL_START: u8 = 2;
+pub const PHASE_INTERVAL_END: u8 = 3;
